+++ libs/search-widget/src/widgets/search-widget/SearchBar.svelte
@@ -22,9 +22,10 @@
     notEnoughDataMessage,
     widgetFeatures,
     widgetFilters,
-    widgetImageRagStrategies, widgetJsonSchema,
+    widgetImageRagStrategies,
+    widgetJsonSchema,
     widgetPlaceholder,
-    widgetRagStrategies
+    widgetRagStrategies,
   } from '../../core/stores/widget.store';
   import {
     activatePermalinks,
@@ -75,6 +76,7 @@
   export let not_enough_data_message = '';
   export let ask_to_resource = '';
   export let max_tokens: number | undefined = undefined;
+  export let max_paragraphs: number | undefined = undefined;
   export let query_prepend = '';
   export let json_schema = '';
   export let vectorset = '';
@@ -102,7 +104,7 @@
   export function search(query: string, filters?: string[]) {
     searchQuery.set(query);
     if (filters) {
-      searchFilters.set({filters});
+      searchFilters.set({ filters });
     }
     typeAhead.set(query || '');
     triggerSearch.next();
@@ -211,6 +213,7 @@
         generative_model: generativemodel,
         ask_to_resource,
         max_tokens,
+        max_paragraphs,
         query_prepend,
         vectorset,
       },